requirements.txt
Django==5.0.1
python-telegram-bot==21.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
whitenoise==6.6.0
plotly==5.18.0